Update testing stucture to accomodate future Rust script and additional test scenarios
