+++ packages/web-components/fast-btr/src/fixtures/binding/binding.fixture.html
@@ -11,6 +11,6 @@
         <f-template name="custom-element">
             <template>{{text}}</template>
         </f-template>
-        <script type="module" src="/main.js"></script>
+        <script type="module" src="/binding/main.js"></script>
     </body>
 </html>
