+++ packages/web-components/fast-btr/src/fixtures/binding/binding.html
@@ -1,0 +1,1 @@
+{{text}}
