+++ packages/web-components/fast-btr/src/fixtures/binding/binding.json
