+++ packages/web-components/fast-btr/src/fixtures/binding/binding.spec.ts
