+++ packages/web-components/fast-btr/src/fixtures/binding/main.ts
